+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "dev": "nodemon server.js",
     "start": "node server.js",
-    "test": "jest"
+    "test": "jest --verbose --silent"
   },
   "keywords": [],
   "author": "Gaizka",
@@ -17,6 +17,7 @@
     "nodemon": "^2.0.21"
   },
   "devDependencies": {
+    "axios": "^1.3.4",
     "jest": "^29.5.0"
   }
 }
