+++ server/__test__/getRandom.test.js
@@ -1,4 +1,4 @@
-const getRandomInt = require('../utils/getRandom');
+const {getRandomInt} = require('../utils/getRandom');
 
 test('Get a random integer less than given value', (value = 3) => {
     expect(getRandomInt(value)).toBeGreaterThanOrEqual(0);
