+++ server/__test__/getResult.test.js
@@ -1,4 +1,4 @@
-const getTirada = require('../utils/getResult');
+const {getTirada} = require('../utils/getResult');
 
 test('Get piedra, papel or tijera from a random integer (0, 1, 2)', () => {
     expect(getTirada(0)).toBe("piedra");
