--- server/test/saveData.test.js
@@ -1,28 +1,0 @@
-const { saveData, reset } = require('../utils/saveData');
-let savedData = [
-    {
-        a: "piedra",
-        b: "papel",
-        c: "a"
-    },
-    {
-        a: "papel",
-        b: "papel",
-        c: "c"
-    }
-];
-
-
-test('data is added to main array savedData', () => {
-    //savedData = [data, ...savedData];
-    const data = {
-        a: "piedra",
-        b: "papel",
-        c: "c"
-        };
-    expect(saveData(data)).toContain(data);
-})
-
-test('Saved data are deleted', () => {
-    expect(reset(savedData)).toBe();
-})
