+++ server/utils/getRandom.js
@@ -2,5 +2,5 @@ function getRandomInt(max) {
     return Math.floor(Math.random() * max);
 };
 
-module.exports = getRandomInt;
+module.exports = {getRandomInt};
